+++ test/customize_savename.jl
@@ -25,20 +25,18 @@ DrWatson.allaccess(::Experiment) = (:n, :c, :x, :species)
 Base.string(::Mouse) = "mouse"
 Base.string(::Cat) = "cat"
 
-
-
-@test savename(e1) == "Experiment_1991-04-13_c=10_n=50_x=0.2"
+@test savename(e1) == "Experiment_1991-04-13_c=10.0_n=50_x=0.2"
 @test savename(e2) == savename(e1)
 
 
 # Add the extra type:
 DrWatson.default_allowed(::Experiment) = (Real, String, Species)
 
-@test savename(e1) ≠ "Experiment_1991-04-13_c=10_n=50_x=0.2"
+@test savename(e1) ≠ "Experiment_1991-04-13_c=10.0_n=50_x=0.2"
 @test savename(e2) ≠ savename(e1)
 
-@test savename(e1) == "Experiment_1991-04-13_c=10_n=50_species=mouse_x=0.2"
-@test savename(e2) == "Experiment_1991-04-13_c=10_n=50_species=cat_x=0.2"
+@test savename(e1) == "Experiment_1991-04-13_c=10.0_n=50_species=mouse_x=0.2"
+@test savename(e2) == "Experiment_1991-04-13_c=10.0_n=50_species=cat_x=0.2"
 
 d = struct2dict(e1)
 @test keytype(d) == Symbol
@@ -56,4 +54,4 @@ end
 s = savename(joinpath("path", "to", "data"), e1)
 @test s[1:4] == "path"
 @test occursin("Experiment", s)
-@test s == joinpath("path", "to", "data", "Experiment_1991-04-13_c=10_n=50_species=mouse_x=0.2")
+@test s == joinpath("path", "to", "data", "Experiment_1991-04-13_c=10.0_n=50_species=mouse_x=0.2")